A bunch of fixes and added bh_visualizer to util.plot_surface.
When running with Bohrium, bh_visualizer is used.
When running with NumPy, Matplotlib is used.

+++ benchmarks/rosenbrock/cpp11_omp/src/rosenbrock.cpp
@@ -30,7 +30,7 @@ int main(int argc, char* argv[])
 
     #pragma omp parallel for
     for(int i=0; i<nelements; ++i) {
-        dataset[i] = i/nelements;
+        dataset[i] = i/(double)nelements;
     }
 
     bp.timer_start();                               // Start timer
