+++ benchmarks/rosenbrock/cpp11_seq/src/rosenbrock.cpp
@@ -26,7 +26,7 @@ int main(int argc, char* argv[])
     // Create the pseudo-data
     double* dataset = (double*)malloc(sizeof(double)*nelements);
     for(int i=0; i<nelements; ++i) {
-        dataset[i] = i/nelements;
+        dataset[i] = i/(float)nelements;
     }
 
     bp.timer_start();                               // Start timer
